+++ CHANGELOG.md
@@ -13,6 +13,8 @@
 - Fix exporting zero scale not importing correctly in 3ds Max.
 - Fix window size cutting off text in export window.
 - Fix object name not being undone after undo convert to prefab.
+- Fix settings reset button not working.
+- Fix not being able to export to a new folder on Mac.
 
 ## [4.1.0-pre.2] - 2021-05-05
 ### Known Issues